uint16

+++ jetson/src/sensor_ingestion/ingest_gi.py
@@ -213,7 +213,7 @@ def on_new_thermal_sample(appsink):
         return Gst.FlowReturn.ERROR
 
     try:
-        frame = np.frombuffer(map_info.data, dtype=np.uint8) # format is GRAY16_LE so 16bit
+        frame = np.frombuffer(map_info.data, dtype=np.uint16) # format is GRAY16_LE so 16bit
         frame = frame.reshape((height, width))
         latest_thermal = frame
         update_buffer()
